Preliminary fix for missing window
Problems exist with the way JobHandler runs when the srcbuf can't be
found. Since the problematic code is trying to clean up the srcbuf
(e.g., removing the local mapping to stop the running makejob), the fix
is to call run this code conditioned on the buffer actually being found.